Revert "галерея волшебников"
