+++ wizard-battle/src/components/Battle/Battle.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import styles from './Battle.module.css';
+
+export default function Battle() {
+  return (
+        <>
+        </>
+  );
+}
