--- wizard-battle/src/components/Battle/Battle.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import styles from './Battle.module.css';
-
-export default function Battle() {
-  return (
-        <>
-        </>
-  );
-}
