+++ wizard-battle/src/components/Feedback/Feedback.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import styles from './Feedback.module.css';
+
+export default function Feedback() {
+  return (
+          <>
+          </>
+  );
+}
