--- wizard-battle/src/components/Feedback/Feedback.module.css
@@ -1,35 +1,0 @@
-.feedBack {
-  max-width: 410px;
-  margin: 74px auto 0 auto;
-}
-
-.feedBack__container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.feedBack__err {
-  color: red;
-}
-
-.feedBack__stars {
-  
-}
-
-.feedBack__star {
-  width: 50px;
-  height: 50px;
-}
-
-.feedBack__checkBoxContainer {
-    display: flex;
-
-}
-
-.feedBack__text {
-  margin: 0;
-  padding: 0;
-  font-size: 20px;
-}
