+++ wizard-battle/src/components/Footer/Footer.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import styles from './Footer.module.css';
+
+export default function Footer() {
+  return (
+          <>
+          <section className={styles.footer}>
+            <h1>footer</h1>
+          </section>
+          </>
+  );
+}
