+++ wizard-battle/src/components/Header/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import styles from './Header.module.css';
+
+export default function Header() {
+  return (
+          <>
+            <section className={styles.header}>
+              <h1>header</h1>
+            </section>
+          </>
+  );
+}
