+++ wizard-battle/src/components/Header/Header.module.css
@@ -2,6 +2,4 @@
     width: 100%;
     background-color: pink;
     height: 50px;
-    display: flex;
-    justify-content: space-between;
 }
