+++ wizard-battle/src/components/IndependentSelect/IndependentSelect.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import styles from './IndependentSelect.module.css';
+
+export default function IndependentSelect() {
+  return (
+          <>
+            <section className={styles.header}>
+              <h1>Выбор вручную</h1>
+            </section>
+          </>
+  );
+}
