--- wizard-battle/src/components/IndependentSelect/IndependentSelect.module.css
@@ -1,18 +1,0 @@
-.manual {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.manual__container {
-    width: 1100px;
-    display: flex;
-    justify-content: space-between;
-}
-
-.manual__button {
-    width: 50px;
-    height: 30px;
-    border: 1px solid;
-    margin-top: 50px;
-}
