+++ wizard-battle/src/components/PopupWithMessage/PopupWithMessage.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import styles from './PopupWithMessage.module.css';
+
+export default function PopupWithMessage({ setIsOpenPopup, text, children }) {
+  function onClose() {
+    setIsOpenPopup(false);
+  }
+  return (
+    <>
+    <div onClick={onClose} className={styles.popup__overlay}></div>
+      <div className={styles.popup}>
+        <h2 className={styles.popup__title}>{text}</h2>
+        <button onClick={onClose} className={styles.popup__close}></button>
+      {children}
+    </div>
+    </>
+  );
+}
