+++ wizard-battle/src/components/PopupWithMessage/PopupWithMessage.module.css
@@ -10,7 +10,7 @@
 }
 
 .popup__close {
- width: 100px;
+ width: 50px;
  height: 50px;
  background-color: blue;
 }
